backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
elevenlabs>=0.2.27
httpx>=0.27.0,<1.0.0
pydantic>=2.7.0,<3.0.0
